requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# AI Services
openai==1.12.0
python-dotenv==1.0.0

# Audio Processing
pyaudio==0.2.14
pydub==0.25.1
numpy==1.26.3

# Utilities
aiofiles==23.2.1
httpx==0.26.0
